CB/bp: fix wrong theme class when logged out if default theme is changed (#22408)
If you don't use the `auto` theme as the default, the `<html>` tag has
`theme-auto` as it's class when users are logged out. This PR changes it
to use the correct theme class for the default theme when logged out.

(cherry picked from commit 9ffaf19d89b24589f4f4ec1211881102cafc7370)
Signed-off-by: crystal <crystal@noreply.codeberg.org>