CB/ui: move style overrides to this repo (!53)
Moves the contents of codeberg.css from the Codeberg-Infrastrcuture/build-deploy-forgejo repository to this one, converting them to a less file.

Also adds overrides to swap the Codeberg logo to the blue version when the theme uses a light background for the header and footer.

fixes https://codeberg.org/Codeberg/Community/issues/880

Co-authored-by: crystal <71373843+CrystalCommunication@users.noreply.github.com>
Reviewed-on: https://codeberg.org/Codeberg/forgejo/pulls/53
Co-committed-by: crystal <crystal@noreply.codeberg.org>