CB/bp: Fix api error message if fork exists
The service returns a custom error message which is not tracked here.

Change order of options

The case that the fork already exists should be more common, followed by
the case that a repo with the same name already exists for other
reasons.

The case that the global repo limit is hit is probably not the
likeliest.