CB/fix: Log SearchIssues error
`ctx.Error` is being used in the context of web route, which only cares
about the first argument as title and second argument as content, any
additional arguments are not used. Notably this function doesn't do any
logging. Temporary workaround this by doing the logging for this
specific function error (which is causing 500 errors for
`/issues/search` endpoint).