CB/fix: Archive checksums
Due to a mistake, all current checksums were generated with the legacy
implementation (external gzip) until this patch.
Old archives were also generated this way.

This ensures that all old archives will be generated with this method
until a time in the future, where we want to rely on Git's defaults for
archive generation